Move react package from services to utils
