+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/react/buildHocDisplayName.js
