+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/react/index.js
