+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/react/tests/buildHocDisplayName.test.js
